Change variable types to long long int for precision

+++ F.Digits Summation.cpp
@@ -3,10 +3,10 @@
 using namespace std;
  
 int main() {
-  int N,M;
+  long long int N,M;
   cin>>N>>M;
-  int n1 = N % 10;
-  int n2 = M % 10;
+  long long int n1 = N % 10;
+  long long int n2 = M % 10;
   cout<<n1+n2; 
   return 0;
 }
